+++ updater/__init__.py
@@ -1,0 +1,3 @@
+from updater.update_service import UpdateService
+
+__all__ = ["UpdateService"]